termina item 5

+++ Semana 2/ex2/fancy_arithmetic_mean.py
@@ -18,41 +18,41 @@ def fancy_arithmetic_mean():
     # The "format" method is used to better organize the code.
     elif number_of_numbers == 1:
         print("Please enter the numbers, one in each line ")
-        number_1 = float(input())
+        number_1 = int(input())
         mean = number_1
         print("The arithmetic mean of the numbers is {}".format(mean))
 
     elif number_of_numbers == 2:
         print("Please enter the numbers, one in each line ")
-        number_1 = float(input())
-        number_2 = float(input())
+        number_1 = int(input())
+        number_2 = int(input())
         mean = (number_1 + number_2) / 2
         print("The arithmetic mean of the numbers is {}".format(mean))
 
     elif number_of_numbers == 3:
         print("Please enter the numbers, one in each line ")
-        number_1 = float(input())
-        number_2 = float(input())
-        number_3 = float(input())
+        number_1 = int(input())
+        number_2 = int(input())
+        number_3 = int(input())
         mean = (number_1 + number_2 + number_3) / 3
         print("The arithmetic mean of the numbers is {}".format(mean))
 
     elif number_of_numbers == 4:
         print("Please enter the numbers, one in each line ")
-        number_1 = float(input())
-        number_2 = float(input())
-        number_3 = float(input())
-        number_4 = float(input())
+        number_1 = int(input())
+        number_2 = int(input())
+        number_3 = int(input())
+        number_4 = int(input())
         mean = (number_1 + number_2 + number_3 + number_4) / 4
         print("The arithmetic mean of the numbers is {}".format(mean))
 
     elif number_of_numbers == 5:
         print("Please enter the numbers, one in each line ")
-        number_1 = float(input())
-        number_2 = float(input())
-        number_3 = float(input())
-        number_4 = float(input())
-        number_5 = float(input())
+        number_1 = int(input())
+        number_2 = int(input())
+        number_3 = int(input())
+        number_4 = int(input())
+        number_5 = int(input())
         mean = (number_1 + number_2 + number_3 + number_4 + number_5) / 5
         print("The arithmetic mean of the numbers is {}".format(mean))
